requirements.txt
google-api-python-client==2.47.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
pypandoc==1.12
requests==2.32.4
pytube==15.0.0
yt-dlp==2024.9.27

gspread==5.12.0
jinja2==3.1.2